Invitation work and Change directory
Invitation work and Change directory

+++ client/src/components/Dashboard/MyBuddies/GroupOptions.jsx
@@ -1,7 +1,7 @@
 import { Component } from "react";
 
 // Api actions
-import { deleteGroup } from "../../../actions/groupService";
+import { deleteGroup, leaveGroup } from "../../../actions/groupService";
 
 import AddEditGroup from "./AddEditGroup";
 import InviteBuddies from "./InviteBuddies";
@@ -14,6 +14,7 @@ class GroupOptions extends Component {
         // This binding is necessary to make `this` work in the callback
         this.toggleEditGroupModal = this.toggleEditGroupModal.bind(this, props);
         this.delete = this.delete.bind(this, props);
+        this.leave = this.leave.bind(this, props);
 
         this.state = {
             currentUser: JSON.parse(localStorage.getItem("user")) || {},
@@ -25,15 +26,15 @@ class GroupOptions extends Component {
     }
     
     render() {
-        console.log("It loaded me again too");
         return (
             <div className="container">
                 <div hidden={this.state.renaming || this.state.inviting} className="modal-container group-options-modal z-50 p-3 drop-shadow-lg rounded-md absolute left-25 top-5 mt-2 table w-[150px] md:w-[150px]">
                     <div className="modal-title">{this.state.group.name}</div>
                     <ul>
-                        <li onClick={() => this.setState(old => ({ renaming: !old.renaming }))} >Rename</li>
-                        <li onClick={() => this.setState(old => ({ inviting: !old.inviting }))}>Add Members</li>
-                        <li onClick={() => this.delete()}>Delete</li>
+                        {this.state.group?.createdBy !== this.state.currentUser?._id && <li onClick={() => this.leave()}>Leave</li>}
+                        {this.state.group?.createdBy === this.state.currentUser?._id && <li onClick={() => this.setState(old => ({ renaming: !old.renaming }))} >Rename</li>}
+                        {this.state.group?.createdBy === this.state.currentUser?._id &&<li onClick={() => this.setState(old => ({ inviting: !old.inviting }))}>Add Members</li>}
+                        {this.state.group?.createdBy === this.state.currentUser?._id &&<li onClick={() => this.delete()}>Delete</li> }
                     </ul>
                 </div>
                 {this.state.renaming && <AddEditGroup group={this.state.group} isEdit="true" closeModal={this.toggleEditGroupModal} />}
@@ -47,8 +48,12 @@ class GroupOptions extends Component {
         props.close(props.group);
     }
 
+    // Leave group
+    async leave(props) {
+        await leaveGroup(this.state.group._id, this.state.currentUser?._id);
+        props.close(props.group);
+    }
     
-
     // Delete current group.
     async delete(props) {
         await deleteGroup(this.state.group._id);
